Move BPMNCellStyle to a dedicated file
